Assessment3_CV_Project/requirements.txt
# Core application
streamlit
numpy
pandas
opencv-python
Pillow

# Model training / inference
ultralytics

# Verification
pytest
